seperate bot and web
