+++ bot/commands/adminonly.command.js
@@ -1,4 +1,4 @@
-import { role } from '../config.js';
+import { role } from '../../config.js';
 
 const adminonly = {
   name: 'adminonly',
