+++ bot/commands/hacktrack.command.js
