+++ bot/commands/issue.command.js
