+++ web/index.js
@@ -1,0 +1,14 @@
+import express from 'express';
+import path from 'path';
+const __dirname = path.resolve();
+const pages_directory = '/web/pages/';
+
+const app = express();
+
+app.get('/', (req, res) => {
+  res.sendFile(path.join(__dirname, pages_directory, 'index.html'));
+});
+
+export async function start_web() {
+  app.listen(3000, () => console.log('Hackerspace MMU web server started'));
+}
